Moves crypto algorithms wrapping to separated directory
