+++ src/lib/Makefile.am
@@ -12,15 +12,18 @@ endif
 
 dist_man_MANS		= librnp.3
 librnp_la_SOURCES	= \
-	bn.c \
+	crypto/bn.c \
+	crypto/dsa.c \
+	crypto/ecdh.c \
+	crypto/ecdsa.c \
+	crypto/eddsa.c \
+	crypto/elgamal.c \
+	crypto/rsa.c \
+	crypto/s2k.c \
+	crypto/sm2.c \
 	bufgap.c \
 	compress.c \
 	crypto.c \
-	dsa.c \
-	ecdh.c \
-	ecdsa.c \
-	eddsa.c \
-	elgamal.c \
 	generate-key.c \
 	hash.c \
 	list.c \
@@ -33,10 +36,7 @@ librnp_la_SOURCES	= \
 	pgp-key.c \
 	reader.c \
 	rnp.c \
-	rsa.c \
-	s2k.c \
 	signature.c \
-	sm2.c \
 	symmetric.c \
 	validate.c \
 	writer.c
