+++ src/lib/crypto/ecdsa.c
