+++ src/lib/crypto/elgamal.h
