+++ src/lib/validate.c
@@ -76,9 +76,10 @@ __RCSID("$NetBSD: validate.c,v 1.44 2012/03/05 02:20:18 christos Exp $");
 #endif
 
 #include "packet-parse.h"
+#include <rnp/rnp_sdk.h>
+
 #include "packet-show.h"
 #include "signature.h"
-#include <rnp/rnp_sdk.h>
 #include "readerwriter.h"
 #include "utils.h"
 #include "memory.h"
